refactor(arch): v0.4 architecture skeleton + Hi-C migration (PR 1/5)
This is the first PR in a 5-PR refactor that aligns epione's
subpackage structure with omicverse's hybrid axis convention:

  epione.bulk.{atac, hic}      modality × scale specific
  epione.single.{atac, hic}    modality × scale specific
  epione.pp / .tl / .pl        cross-modality
  epione.io                    pure format read/write
  epione.upstream              FASTQ → BAM → bigwig / pairs → cool pipelines
  epione.core                  pure-Python infrastructure
  epione.data                  reference registries (Genome, PWM DB)
  epione.datasets              fetchable example datasets

PR 1 lands the skeleton for all 8 new subpackages and migrates the
Hi-C code (the smallest blast-radius modality):

  epione/sc_hic/            → epione/single/hic/
                              (preserves flat API: load_scool_cells,
                               impute_cells, embedding, plot_*)
  epione/hic/correct.py     → epione/bulk/hic/_balance.py
  epione/hic/plot.py        → epione/bulk/hic/_plot.py
  epione/hic/build.py       → epione/upstream/_pairs.py
                              (pairs_from_bam / pairs_to_cool)

Backwards compatibility:

  - epione.hic and epione.sc_hic are now re-export shims that emit
    DeprecationWarning. Old import paths still resolve to the new
    code (verified via id-equality on function objects).
  - .gitignore tightened so the `data` rule no longer hides the new
    epione/data/ subpackage.

Tutorials + tests updated:

  - tests/test_hic.py        → tests/test_bulk_hic.py
  - tests/test_sc_hic.py     → tests/test_single_hic.py
  - epione_guide/tutorials/{bulk,single}/t_*hic_basics.ipynb point at
    the new paths (epi.bulk.hic / epi.single.hic / epi.upstream).

PR 2 will migrate epione.align → epione.upstream + bulk/_*.py →
bulk/atac/ + single/_*.py → single/atac/. PR 3 unifies plotting
into epione.pl by chart type. PR 4 carves utils/ into core/ + io/
and seeds datasets/ with Nagano / Drosophila / TOBIAS fetchers.
PR 5 removes the deprecation aliases.

Tests: 63/63 pass (no functional change; pure structural migration).
