+++ epione/bulk/atac/__init__.py
@@ -1,0 +1,17 @@
+"""Bulk ATAC-seq analysis — counterpart to :mod:`epione.single.atac`.
+
+Future home (during the v0.4 refactor PR 2):
+
+    * Footprint analysis (currently :mod:`epione.bulk._footprint_archr`,
+      :mod:`epione.tl._footprint`, :mod:`epione.tl._bindetect`,
+      :mod:`epione.tl._atacorrect`)
+    * Differential peak analysis (currently
+      :mod:`epione.tl._differential_peaks`)
+    * Peak-to-gene linkage (currently :mod:`epione.tl._peak2gene`)
+
+Empty in PR 1; populated when :mod:`epione.bulk` migrates its
+``_*.py`` files here.
+"""
+from __future__ import annotations
+
+__all__: list[str] = []
